BitString Problems

BitString data structure

The fitness functions for BitString problems just accept BitArray or Array{Bool,1} as the types of solution instances. However, it is recommended to use BitArray instead of Array{Bool,1} since BitArray uses memory more efficiently. Nevertheless, solution instances with type Array{Bool,1} are changed to BitArray automatically.

Built-in problems

Pidoh.OneMax — Type

struct OneMax <: AbstractBitstringProblem
    n::Integer
end

Maximization of the number of true in a BitArray. The mathematical definition of this problem is maximizing the number of one-bits in a bitstring, which can be seen in the following formula:

\[OneMax(x_1,\dots,x_n)\coloneqq\sum_{i=1}^{n}x_i\]

Examples

julia> problem = OneMax(4)
OneMax(4)

julia> fitness(BitArray([true,true,false,true]), problem)
3

julia> optimum(problem)
4