Simulated Annealing and Metropolis Algorithms

Examples

using Pidoh
algorithm = SimulatedAnnealing(FixedCooling(10.0), stop = FixedBudget(10^6));
instance = Instance(generator = RandBitStringIP(100), OneMax(100));
experiment = Experiment("SAonOneMax", [algorithm], [instance], save = false);
res = run(experiment);
last(res.jobs[1].trace.rows, 10)

iteration	fitness

Int64	Int64

10 rows × 2 columns

1	229842	95

2	229854	96

3	229869	95

4	229876	96

5	229879	95

6	229880	96

7	229885	97

8	229894	98

9	229896	99

10	229898	100

Built-in algorithms

Pidoh.SimulatedAnnealing — Type

struct SimulatedAnnealing <: AbstractSimulatedAnnealing
    cooling::AbstractCooling
    stop::AbstractStop
    name::LaTeXString
    temperature::Float64
end

Built-in cooling schedules

Fixed Cooling Schedulers

Pidoh.FixedCooling — Type

struct FixedCooling <: AbstractCooling
    temperature::Float64
end

Using this cooling scheduler, the algorithm accepts a worser solution with probability $\alpha^{-\Delta}$, where $\alpha$ is the parameter temperature and $\Delta$ is the absolute difference of the fitness functions between the parent and child.

Through FixedCooling, you are basicaully using Metropolis algorithm.

Dynamic Cooling Schedulers

There is no dynamic cooling schedulers implemented yet.